alguns retoques no lzw

+++ source/LZW2/LZW2.h
@@ -28,7 +28,7 @@ private:
     std::map<std::vector<uint8_t>, lzw_code_t> compression_map;
     std::map<lzw_code_t, std::vector<uint8_t>> decompression_map;
 
-    long int max_sequences = (1 << sizeof(lzw_code_t)*8) - 1; // A princípio pode ser no máximo o que lzw_code_t puder comportar
+    unsigned long long max_sequences = (1 << sizeof(lzw_code_t)*8) - 1; // A princípio pode ser no máximo o que lzw_code_t puder comportar
 
     bool restart_map_on_overflow = true;
